+++ Employee Management System/route/employee_route.py
@@ -1,13 +1,19 @@
 
-from fastapi import APIRouter, HTTPException, Query
+from fastapi import APIRouter, HTTPException, Query, Depends
 from config.databases import sql, cursor
 from model.employee_models import Register
 from schema.employee_schema import list_serial
+from .auth_route import authenticate_user, Role
+
 
 employee_router = APIRouter()
 
 @employee_router.post("/")
-async def enter_employee_details(info: Register):
+async def enter_employee_details(info: Register, current_user: dict = Depends(authenticate_user)):
+    # Check if the user is admin
+    if current_user["role"] != Role.admin:
+        raise HTTPException(status_code=403, detail="Only admin can add new employees")
+
     sql_query = "INSERT INTO employee VALUES (%s, %s, %s, %s, %s);"
     try:
         cursor.execute(sql_query, (info.employeeId, info.email, info.name, info.salary, info.role))
@@ -18,7 +24,11 @@ async def enter_employee_details(info: Register):
         return {"message": "Record added successfully"}
 
 @employee_router.get("/my_info")
-async def my_info():
+async def my_info(current_user: dict = Depends(authenticate_user)):
+    # Check if the user is admin or manager
+    if current_user["role"] not in [Role.admin, Role.manager, Role.user]:
+        raise HTTPException(status_code=403, detail="Only admin or manager or employee can access employee info")
+
     sql_query = "SELECT * FROM employee;"
     try:
         cursor.execute(sql_query)
@@ -30,7 +40,11 @@ async def my_info():
         return result
 
 @employee_router.put("/{employeeId}")
-async def update_name(employeeId: str, name: str):
+async def update_name(employeeId: str, name: str, current_user: dict = Depends(authenticate_user)):
+    # Check if the user is admin
+    if current_user["role"] != Role.admin:
+        raise HTTPException(status_code=403, detail="Only admin can update employee details")
+
     sql_query = "UPDATE employee SET name = %s WHERE employeeId = %s;"
     try:
         cursor.execute(sql_query, (name, employeeId))
@@ -41,7 +55,11 @@ async def update_name(employeeId: str, name: str):
         return {"message": "Record updated successfully"}
 
 @employee_router.delete("/")
-async def delete_record(employeeId: str = Query(..., description="Employee ID")):
+async def delete_record(employeeId: str = Query(..., description="Employee ID"), current_user: dict = Depends(authenticate_user)):
+    # Check if the user is admin
+    if current_user["role"] != Role.admin:
+        raise HTTPException(status_code=403, detail="Only admin can delete employee records")
+
     sql_query = "DELETE FROM employee WHERE employeeId = %s;"
     try:
         cursor.execute(sql_query, (employeeId,))
